% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtransition-main.R
\docType{package}
\name{Gtransition-package}
\alias{Gtransition-package}
\alias{Gtrasition-package}
\alias{Gtrasition}
\title{Título de lo que hace.}
\description{
Decsripcion de lo que hace
}
\details{
Package: Gtrasition

Type: Package

Detalles de lo que hace.

In the Gtrasition...
}
\examples{
#See examples for functions.
}
\references{
Luquin-Covarrubias M., Morales-Bojorquez E. (2020). Effects of stochastic growth on population dynamics and management quantities estimated from an integrated catch-at-length assessment model: Panopea globosa as case study. Ecologial Modeling 440, 109384. https://doi.org/10.1016/j.ecolmodel.2020.109384

Sullivan P.J., Lai H., Galluci V.F. (1990). A Catch-at-Length analysis that incorporates a stochastic model of growth. Can. J. Fish. Aquat. Sci. 47: 184-198.
}
\author{
Arely Ornella Vargas <aornelasv@ipn.mx>

Josymar Torrejon-Magallanes <ejosymart@gmail.com>
}
\concept{Gtransition}
\concept{distributions}
\concept{growth}
\concept{transition}
