% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtransition-main.R
\name{plot.Mtransition}
\alias{plot.Mtransition}
\title{Plot method for Mtransition class}
\usage{
\method{plot}{Mtransition}(
  x,
  xlab = "X-Text",
  ylab = "Y-Text",
  col = "grey45",
  sizeAxis1 = 0.85,
  sizeAxis2 = 0.5,
  adjY = -15.5,
  ...
)
}
\arguments{
\item{x}{object of class 'Mtransition'.}

\item{xlab}{a title for the x axis.}

\item{ylab}{a title for the y axis.}

\item{col}{color for the barplot.}

\item{sizeAxis1}{a number for the axis X.}

\item{sizeAxis2}{a number for the axis Y.}

\item{adjY}{a number for y label position}

\item{\dots}{Additional arguments to the plot method.}
}
\description{
Plot method for Mtransition class
}
\examples{
output <- mgi(lowerL = 78, upperL = 202, classL = 4, Linf = 197.42, K = 0.1938, method = "vonB")
delta <- output$delta

mat <- transitionM(lowerL = 78, upperL = 202, classL = 4, distribution = "gamma", 
delta = delta, beta = 0.105, sigma = NULL)

plot(mat)
}
