% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtransition-main.R
\name{mgi}
\alias{mgi}
\title{Mean growth increment}
\usage{
mgi(lowerL, upperL, classL, Linf, K, gm = 1, dl = 0.1, method = "vonB")
}
\arguments{
\item{lowerL}{a numeric value that represents...}

\item{upperL}{a numeric value that represents...}

\item{classL}{a numeric value ...}

\item{Linf}{a numeric value ....}

\item{K}{a numeric value ..}

\item{gm}{a numeric value .. \code{gm = 1} as default value??.}

\item{dl}{a numeric value ...  \code{dl = 0.1} as default value??.}

\item{method}{a character string defining the growth equation to be used.}
}
\value{
A list of class 'Gincrement'.

\code{delta} the mean growth increment

\code{Laverage} the average length.
}
\description{
Estimate mean growth increment, for the individuals in length class I is then the average change in length of individuals initially in length class
}
\details{
Estimate mean growth increment
}
\examples{
output <- mgi(lowerL = 78, upperL = 202, classL = 4, Linf = 197.42, K = 0.1938, method = "vonB")

output
output$delta
output$Laverage
}
\references{
Sullivan P.J., Lai H., Galluci V.F. (1990). A Catch-at-Length analysis that incorporates a stochastic model of growth. Can. J. Fish. Aquat. Sci. 47: 184-198.
}
