% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtransition-main.R
\name{plot.Mtransition}
\alias{plot.Mtransition}
\title{Plot method for Mtransition class}
\usage{
\method{plot}{Mtransition}(
  x,
  units = "mm.",
  xlab = "Length class",
  ylab = "Y-Text",
  col = "grey45",
  sizeAxis1 = 0.5,
  sizeAxis2 = 0.5,
  adjY = -20,
  filename = "myplot",
  savePDF = FALSE,
  widthPDF = 3,
  heightPDF = 10,
  savePNG = FALSE,
  widthPNG = 300,
  heightPNG = 1000,
  resPNG = 110,
  ...
)
}
\arguments{
\item{x}{object of class 'Mtransition'.}

\item{units}{a character string related to the measure of length.}

\item{xlab}{a title for the x axis.}

\item{ylab}{a title for the y axis.}

\item{col}{color for the barplot.}

\item{sizeAxis1}{a number for the axis X.}

\item{sizeAxis2}{a number for the axis Y.}

\item{adjY}{a number for y label position .}

\item{filename}{filename.}

\item{savePDF}{to save or not the plot in pdf format.}

\item{widthPDF}{width resolution for pdf format.}

\item{heightPDF}{height resolution for pdf format.}

\item{savePNG}{to save or not the plot in png format.}

\item{widthPNG}{width resolution for png format.}

\item{heightPNG}{height resolution for png format.}

\item{resPNG}{png resolution for png format.}

\item{\dots}{Additional arguments to the plot method.}
}
\description{
Plot method for Mtransition class
}
\examples{
output <- mgi(lowerL = 78, upperL = 202, classL = 4, Linf = 197.42, k = 0.1938, method = "vonB")
delta <- output$delta

mat <- transitionM(lowerL = 78, upperL = 202, classL = 4, distribution = "gamma", 
delta = delta, beta = 0.105, sigma = NULL)

plot(mat)
}
