% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtransition-main.R
\name{transitionM}
\alias{transitionM}
\title{Transition Matrix}
\usage{
transitionM(
  lowerL,
  upperL,
  classL,
  distribution = "gamma",
  delta,
  beta = NULL,
  sigma = NULL
)
}
\arguments{
\item{lowerL}{a numeric value that represents...}

\item{upperL}{a numeric value that represents...}

\item{classL}{a numeric value ...}

\item{distribution}{a character string defining the growth equation to be used.}

\item{delta}{a numeric vector...}

\item{beta}{a numeric value...}

\item{sigma}{a numeric value...}
}
\value{
A list of class 'Mtransition'.

\code{mcdf} the mean growth increment

\code{G} the average length.
}
\description{
Estimate .....
}
\details{
Estimate .....
}
\references{
Luquin-Covarrubias, M.A. and E. Morales-Bojórquez. 2021. Effects of stochastic growth on population dynamics and management quantities estimated from an integrated catch-at-length assessment model: Panopea globosa as case study. Ecological Modelling. 440: 109384. https://doi.org/10.1016/j.ecolmodel.2020.109384.
output <- mgi(lowerL = 78, upperL = 202, classL = 4, Linf = 197.42, K = 0.1938, method = "vonB")
delta <- output$delta

mat <- transitionM(lowerL = 78, upperL = 202, classL = 4, distribution = "gamma", 
delta = delta, beta = 0.105, sigma = NULL)

mat
}
