% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtransition-main.R
\name{transitionM}
\alias{transitionM}
\title{Transition Matrix}
\usage{
transitionM(
  lowerL,
  classL,
  Linf,
  distribution = "gamma",
  delta,
  beta = NULL,
  sigma = NULL
)
}
\arguments{
\item{lowerL}{a numeric value that represents...}

\item{classL}{a numeric value ...}

\item{Linf}{a numeric value ....}

\item{distribution}{a character string defining the growth equation to be used.}

\item{delta}{a numeric vector...}

\item{beta}{a numeric value...}

\item{sigma}{a numeric value...}
}
\value{
A list of class 'mtransition'.

\code{mcdf} the mean growth increment

\code{G} the average length.
}
\description{
Estimate .....
}
\details{
Estimate .....
}
\examples{
output <- mgi(lowerL = 10, classL = 10, Linf = 60, K = 0.3, method = "vonB")
delta <- output$delta

mat <- transitionM(lowerL = 10, classL = 10, Linf = 60, distribution = "gamma", 
delta = delta, beta = 1.5, sigma = NULL)

mat
}
\references{
Sullivan P.J., Lai H., Galluci V.F. (1990). A Catch-at-Length analysis that incorporates a stochastic model of growth. Can. J. Fish. Aquat. Sci. 47: 184-198.
}
